feat: updateBranch - Criação de um novo form dentro de list.ejs, sua estilização, criação da rota delete e implementação de remover tasks no banco de dados apenas clicando no checkbox

+++ date.js
@@ -1,7 +1,7 @@
 
 exports.getDate = function () {
 
-    const options = { weekday: 'short', month: 'long', day: 'numeric', year: 'numeric'};
+    const options = { weekday: 'short', month: 'short', day: 'numeric', year: 'numeric'};
     
     const today = new Date();
